refactor(project): restructure project imports and remove unnecessary `__init__.py` files
