--- devtools_release_notifier/notifiers/__init__.py
@@ -1,5 +1,0 @@
-"""Notifiers for various platforms."""
-
-from devtools_release_notifier.notifiers.discord import DiscordNotifier
-
-__all__ = ["DiscordNotifier"]
