--- devtools_release_notifier/sources/__init__.py
@@ -1,13 +1,0 @@
-"""Release information sources."""
-
-from devtools_release_notifier.sources.base import ReleaseSource
-from devtools_release_notifier.sources.github_commits import GitHubCommitsSource
-from devtools_release_notifier.sources.github_releases import GitHubReleaseSource
-from devtools_release_notifier.sources.homebrew_cask import HomebrewCaskSource
-
-__all__ = [
-    "ReleaseSource",
-    "GitHubReleaseSource",
-    "HomebrewCaskSource",
-    "GitHubCommitsSource",
-]
